feat(driver): make `DriverType` available on all platform
